requirements.txt
paho-mqtt 
gpiozero
